requirements.txt
numpy
pandas
scikit-learn
opencv-python
pillow
flask
